feat: add share dialog
